+++ src/features/editor/extensions/embed/index.ts
@@ -1,6 +1,6 @@
 import { Node } from "@tiptap/react";
 import { getEmbedTypeFromElement } from "@/features/editor/lib/embed";
-import { generateEmbedOutputSpec } from "./generate-embed-outputSpec";
+import { generateEmbedOutputSpec } from "./generate-embed-output-spec";
 
 export const Embed = Node.create({
   name: "embed",
